requirement.txt
numpy
pandas
scikit-learn
opencv-python
Pillow
matplotlib
scipy
tk
